Fixing a few build errors that slipped through last commit/push.

+++ basestation/src/anki/cozmo/basestation/behaviors/behaviorFollowMotion.cpp
@@ -108,7 +108,7 @@ IBehavior::Status BehaviorFollowMotion::UpdateInternal(Robot& robot, double curr
 
 Result BehaviorFollowMotion::InterruptInternal(Robot& robot, double currentTime_sec, bool isShortInterrupt)
 {
-  _actionRunning = ActionConstants::INVALID_TAG;
+  _actionRunning = (u32)ActionConstants::INVALID_TAG;
   _lastInterruptTime_sec = currentTime_sec;
   _holdHeadDownUntil = -1.0f;
